DMS_0013: Fixing double data type

+++ src/main/java/ctu/cit/se/dam_service/dtos/responses/damschedules/RetrieveDamScheduleBySelectedDateResDTO.java
@@ -28,7 +28,7 @@ public class RetrieveDamScheduleBySelectedDateResDTO {
     @JsonProperty("damHeight")
     private Double height;
     @JsonProperty("damCapacity")
-    private Integer capacity;
+    private Double capacity;
     @JsonProperty("damRiverName")
     private String riverName;
     @JsonProperty("damSchedulesAtThisDate")
